Tidy up yamls in calicoctl/test-data and improve test_convert.py
Remove yamls that were not used by any test and organize them by test
(test_convert.py and multi_context-test.go), plus have inputs and
expected outputs for test_convert.py, which now compares them with
the conversion output.

+++ calicoctl/tests/st/calicoctl/test_convert.py
@@ -14,6 +14,7 @@
 import json
 import logging
 import copy
+import os
 
 from nose_parameterized import parameterized
 
@@ -25,20 +26,23 @@ logging.basicConfig(level=logging.DEBUG, format="%(message)s")
 logger = logging.getLogger(__name__)
 
 convert_files = [
-        ("test-data/v1/bgppeer-global.yaml",),
-        ("test-data/v1/bgppeer-node.yaml",),
-        ("test-data/v1/bgppeer-node2.yaml",),
-        ("test-data/v1/multi-resource.yaml",),
-        ("test-data/v1/node.yaml",),
-        ("test-data/v1/test3.yaml",),
-        ("test-data/v1/migration/bgppeer.yaml",),
-        ("test-data/v1/migration/hostendpoint.yaml",),
-        ("test-data/v1/migration/ippool.yaml",),
-        ("test-data/v1/migration/node.yaml",),
-        ("test-data/v1/migration/policy.yaml",),
-        ("test-data/v1/migration/profile.yaml",),
-        ("test-data/v1/migration/workloadendpoint.yaml",),
-        ("test-data/k8s/k8s-networkpolicy.yaml",),
+        ("test-data/convert/input/v1/bgppeer-global.yaml",),
+        ("test-data/convert/input/v1/bgppeer-node.yaml",),
+        ("test-data/convert/input/v1/bgppeer-node2.yaml",),
+        ("test-data/convert/input/v1/multi-resource.yaml",),
+        ("test-data/convert/input/v1/node.yaml",),
+        ("test-data/convert/input/v1/test3.yaml",),
+        ("test-data/convert/input/v1/migration/bgppeer.yaml",),
+        ("test-data/convert/input/v1/migration/hostendpoint.yaml",),
+        ("test-data/convert/input/v1/migration/ippool.yaml",),
+        ("test-data/convert/input/v1/migration/node.yaml",),
+        ("test-data/convert/input/v1/migration/policy.yaml",),
+        ("test-data/convert/input/v1/migration/profile.yaml",),
+        ("test-data/convert/input/v1/migration/workloadendpoint.yaml",),
+        ("test-data/convert/input/k8s/k8s-networkpolicy.yaml",),
+        ("test-data/convert/input/k8s/k8s-networkpolicy-invalid.yaml",),
+        ("test-data/convert/input/k8s/k8s-networkpolicy-multiple.yaml",),
+        ("test-data/convert/input/k8s/k8s-networkpolicy-multiple-invalid.yaml",),
     ]
 
 class TestCalicoctlConvert(TestBase):
@@ -46,32 +50,46 @@ class TestCalicoctlConvert(TestBase):
     Test calicoctl convert
     """
 
-    def _test_convert_successful(self, filename, format="yaml"):
+    def _test_convert(self, filename, format="yaml"):
         """
         Test convert successfully
         """
         # Convert the file
         rc = calicoctl("convert -o %s -f %s" % (format, filename))
-        rc.assert_no_error()
 
-        # With the converted data to a temp file
-        with open("/tmp/converted", 'w') as f:
-            f.write(rc.output)
+        # Get expected conversion output filename (file in output/ dir with
+        # yaml or json extension)
+        output_filename = (filename.replace("input", "output").split(".")[0]
+                           + "." + format)
 
-        # Load the converted data
-        rc = calicoctl("apply -f /tmp/converted")
-        rc.assert_no_error()
+        # If expected conversion output exists then assert that the
+        # conversion is successful, otherwise assert an error occurs
+        if os.path.isfile(output_filename):
+            rc.assert_no_error()
+            with open(output_filename, 'r') as f:
+                expected_output = f.read().rstrip()
+                self.assertEqual(rc.output, expected_output)
+
+            # With the converted data to a temp file
+            with open("/tmp/converted", 'w') as f:
+                f.write(rc.output)
+
+            # Load the converted data
+            rc = calicoctl("apply -f /tmp/converted")
+            rc.assert_no_error()
+        else:
+            rc.assert_error()
 
     @parameterized.expand(convert_files)
     def test_convert_successful_yaml(self, filename):
         """
         Test convert with yaml output.
         """
-        self._test_convert_successful(filename, format="yaml")
+        self._test_convert(filename, format="yaml")
 
     @parameterized.expand(convert_files)
     def test_convert_successful_json(self, filename):
         """
         Test convert with json output.
         """
-        self._test_convert_successful(filename, format="json")
+        self._test_convert(filename, format="json")
